// src/app/modules/checkpoint_module/pages/wizards/checkpoint-firewall/checkpoint-firewall-wizard.interface.ts
import { WizardPost } from '../../../../../pages/wizards/wizards.interface';


// WIZARD POST
export interface CheckpointFirewallWizardPost extends WizardPost {
    authPassword: string
    authProtocol: string
    privacyPassword: string
    privacyProtocol: string
    securityLevel: string
    securityName: string
    snmpCommunity: string
    snmpVersion: string
}